+++ armsrc/mifarecmd.c
@@ -2243,6 +2243,12 @@ void MifareCIdent(void) {
         }
         if (memcmp(buf, "\x0D\x78\x00\x71\x02\x88\x49\xA1\x30\x20\x15\x06\x08\x56\x3D", 15) == 0) {
             isGen = MAGIC_GEN_2;
+            goto OUT;
+        }
+        // test for Ultralight magic gen2
+        if (memcmp(buf, "\x0A\x78\x00\x81\x02\xDB\xA0\xC1\x19\x40\x2A\xB5", 12) == 0) {
+            isGen = MAGIC_GEN_2;
+            goto OUT;
         }
     };